generator/requirements.txt
python-can
paho-mqtt
requests
